requirement.txt
asgiref==3.6.0
Django==3.2.18
future==1.0.0
iso8601==2.1.0
mysqlclient==2.1.0
numpy==1.21.6
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
Pillow==9.5.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
